requirements.txt
alembic==1.7.6
bcrypt==3.2.0
cffi==1.15.0
click==8.0.4
colorama==0.4.4
dnspython==2.2.0
email-validator==1.1.3
Flask==2.0.3
Flask-Admin==1.6.0
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
greenlet==1.1.2
idna==3.3
itsdangerous==2.1.0
Jinja2==3.0.3
Mako==1.1.6
MarkupSafe==2.1.0
Pillow==9.0.1
pycparser==2.21
six==1.16.0
SQLAlchemy==1.4.31
Werkzeug==2.0.3
WTForms==3.0.1
WTForms-SQLAlchemy==0.3
